pyproject.toml
[project]
name = "inventory-management"
version = "0.1.0"
description = ""
authors = [
    {name = "alvaro",email = "alvaro.apariciofdz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "aioredis (>=2.0.1,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "pydantic[email] (>=2.12.4,<3.0.0)"
]

[tool.poetry]
packages = [{include = "inventory_management", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
